Handle broken DWARF data without crash in "Data Types" view
libdwarf will automatically abort when it runs into invalid or
mismatching DWARF data unless we provide a custom error handler.
This adds the custom handler to DwarfInfo and marks the whole
class as invalid if an error occurs while parsing the DWARF data.

Opening a type in Type view which has broken or incomplete DWARF
data leads to crashes all over the place (due to unchecked null
pointers), so the best solution is to simply omit invalid types
from the model and just show a warning that the tree is incomplete.

REVIEW: 126263