CMake: improve binutils detection
On FreeBSD, <demangle.h> lives in /usr/local/include/libiberty/demangle.h,
which is not in the standard search path. That made all the version-
checks fail here (except the one for dis-asm.h, which .. look,
packaging decisions made in the distant past make this stupid now;
just roll with it).