Fix init_disassemble_info configure check
The configure check failed for me, despite having binutils 2.43.1, which
has the signature with the four parameters, leading to a build error
later on, because the three parameter version is used.

Turns out it fails due to bfd.h being "special" and insisting on
including config.h from binutils's source dir:

/usr/x86_64-pc-linux-gnu/include/bfd.h:35:2: error: #error config.h must be included before this header
   35 | #error config.h must be included before this header

This #error is skipped if PACKAGE and PACKAGE_VERSION are defined
though, so just define them.